requirements.txt
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Deprecated==1.2.18
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pycparser==2.22
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
requests==2.32.3
smmap==5.0.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
